requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
Django==4.2.1
django-bootstrap5==23.3
django-crispy-forms==2.0
django-filter==23.2
django-tables2==2.5.3
kombu==5.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
vine==5.0.0
wcwidth==0.2.6
